Split pybind helper into generic and robot-specific part
The bindings of RobotData, RobotBackend, etc. can be done in a
completely generic way, given that action and observation types are
provided via template parameters.

The "N-joint" action and observation provided in this package are
specific to the BLMC CAN robots, though, and do not necessarily fit for
other robots.

Therefore separate the binding code for the two into different helper
functions such that the generic part can be used with other
action/observation types as well.

BREAKING CHANGE: The `create_python_bindings()` function (which includes
the N-joint action/observation) is moved to `pybind_finger.hpp` and
renamed to `create_blmc_can_robot_python_bindings()`.
