+++ srcpy/py_trifinger_types.cpp
@@ -20,11 +20,11 @@
  * \brief Create bindings for TriFinger robot types
  */
 #include <robot_interfaces/finger_types.hpp>
-#include <robot_interfaces/pybind_helper.hpp>
+#include <robot_interfaces/pybind_finger.hpp>
 
 using namespace robot_interfaces;
 
 PYBIND11_MODULE(py_trifinger_types, m)
 {
-    create_python_bindings<TriFingerTypes>(m);
+    create_blmc_can_robot_python_bindings<TriFingerTypes>(m);
 }
